move serve_{initrd, ipxe, kernel} to a boot module
